Hackathon/requirements.txt
streamlit
pymupdf
pillow
requests
msal
pandas
jw-client